pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
replit = "^2.0.0"
apscheduler = "^3.6.3"
psutil = "^5.8.0"
discord = "^1.0.1"
xboxapi = "^2.0.1"
youtube-dl = "^2021.1.24"
aiocron = "^1.3"
mcstatus = "^5.1.1"
pil-compat = "^1.0.0"
gtts = "^2.2.2"
mccq = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
